Centralized duplicate intermediate saving code
